move link-checking logic to a separate module
